third_party/blink/web_tests/external/wpt/css/css-viewport/zoom/computed-initial-expected.txt

This is a testharness.js-based test.
Found 6 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] column-rule-width
  assert_equals: expected "3px" but got "0.3px"
[FAIL] row-rule-width
  assert_equals: expected "3px" but got "0.3px"
[FAIL] column-rule-width: initial;
  assert_equals: expected "3px" but got "0.3px"
[FAIL] row-rule-width: initial;
  assert_equals: expected "3px" but got "0.3px"
[FAIL] column-rule-width (all: initial)
  assert_equals: expected "3px" but got "0.3px"
[FAIL] row-rule-width (all: initial)
  assert_equals: expected "3px" but got "0.3px"
Harness: the test ran to completion.
